+++ sancho-src/src/sancho/view/statusline/LinkEntry$1.java
@@ -1,5 +1,6 @@
 package sancho.view.statusline;
 
+import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.KeyAdapter;
 import org.eclipse.swt.events.KeyEvent;
 import org.eclipse.swt.widgets.Text;
@@ -16,7 +17,7 @@ class LinkEntry$1 extends KeyAdapter {
    }
 
    public void keyPressed(KeyEvent var1) {
-      if ((var1.stateMask & 262144) != 0 && (var1.character == '\n' || var1.character == '\r' || var1.character == 16777296)) {
+      if ((var1.stateMask & SWT.MOD1) != 0 && (var1.character == '\n' || var1.character == '\r' || var1.character == 16777296)) {
          this.this$0.enterLinks(this.val$linkEntryText);
          var1.doit = false;
       }