+++ sancho-src/src/sancho/view/viewer/table/GTableMenuListener$2.java
@@ -1,5 +1,6 @@
 package sancho.view.viewer.table;
 
+import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.KeyAdapter;
 import org.eclipse.swt.events.KeyEvent;
 
@@ -12,9 +13,9 @@ class GTableMenuListener$2 extends KeyAdapter {
    }
 
    public void keyPressed(KeyEvent var1) {
-      if ((var1.stateMask & 262144) != 0 && (var1.stateMask & 131072) != 0 && var1.character == 1) {
+      if ((var1.stateMask & SWT.MOD1) != 0 && (var1.stateMask & 131072) != 0 && var1.character == 1) {
          this.this$0.deselectAll();
-      } else if (var1.stateMask == 262144 && var1.character == 1) {
+      } else if (var1.stateMask == SWT.MOD1 && var1.character == 1) {
          this.this$0.selectAll();
       } else if (var1.keyCode == 127) {
          this.this$0.onDeleteKey();